repo/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.1.0
cryptography==44.0.2
pytest==8.3.5
pytest-flask==1.3.0
pytest-playwright>=0.7.2
playwright>=1.58.0
